Correções

+++ phase_4/app/api/routers/creditCard.py
@@ -12,11 +12,12 @@ access = security.UserHasAccess([UserType.CUSTOMER])
 
 @router.post("/", description="Create a Credit Card")
 def add(
-    creditCardData: FrontToBackEndCreditCardDTO, user: Annotated[User, Depends(access)]
+    creditCardData: FrontToBackEndCreditCardDTO,
+    user: Annotated[User, Depends(access)],
 ):
     success = service.addCreditCard(creditCardData, user.idUsuario)
     if not success:
-        return HTTPException(500, "Não foi possível cadastrar o cartão")
+        raise HTTPException(500, "Não foi possível cadastrar o cartão")
 
 
 @router.get("/", description="Get all customer credit cards")
@@ -27,5 +28,15 @@ def list(user: Annotated[User, Depends(access)]):
 
 @router.delete("/{id}", description="Delete a customer's credit card by its ID")
 def delete(id: int, user: Annotated[User, Depends(access)]):
-    response = service.deleteCreditCardById(id, user.idUsuario)
-    return {"message": response}
+    creditCard = service.getCreditCardByCardId(id)
+
+    if creditCard is None:
+        raise HTTPException(404, "Cartão de crédito não encontrado")
+
+    if creditCard.idClient != user.idUsuario:
+        raise HTTPException(403, "O cartão não pertence ao usuário logado")
+
+    success = service.deleteCreditCardById(id, user.idUsuario)
+
+    if not success:
+        raise HTTPException(500, "Não foi possível remover o cartão")
